Extract BinlogStatement to reduce duplicate logic for MySQL and doris
